[wordle] Run the app in the compatibility mode
It appears that the ~/.local/share/harbour-wordle directory only gets created
in the compability mode. And if that directory doesn't exist (and mapped into
sadbox), then the app state, as well as the history and statistics don't get
preserved between the runs - the app starts from scratch every time, that's a
huge loss of functionality.

Removing X-Sailjail section from the .desktop file seems to be the only
way to turn on that compatibility mode. So be it.